WIP splitting up darc clone
